backend/requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
sentence-transformers
beautifulsoup4
requests
lxml
scrapy
python-dotenv
pydantic
ollama
python-multipart
aiofiles
pandas
numpy
Pillow
tqdm
datasets